+++ CHANGELOG.md
@@ -1443,3 +1443,4 @@ Removed kcal from the macro donut chart entirely since macro-derived calories (p
 - Added an "edited" case to saveFoodData to show a specific snackbar for a successful edit instead of showing "Food logged successfully"
 - Made it so if the user submits the same serving size it just returns
 - Made all dialogs have the same opacity for consistency
+- Added the textStyle dialog code to main.dart instead of having to explicitly write it out every time